+++ pygeoops/__init__.py
@@ -4,8 +4,9 @@ Library with some less common or extended spatial functions.
 
 from pathlib import Path
 
-from pygeoops._difference import *  # noqa: F403
+from pygeoops._buffer_by_m import *  # noqa: F403
 from pygeoops._centerline import *  # noqa: F403
+from pygeoops._difference import *  # noqa: F403
 from pygeoops._extend_line import *  # noqa: F403
 from pygeoops._general import *  # noqa: F403
 from pygeoops._grid import *  # noqa: F403